Adicionado policies para gerenciar a autorização de produtores e consumidores. Corrigido: Deletar consumidor agora também deleta o usuário associado.

+++ app/Http/Controllers/Api/ConsumidorController.php
@@ -8,6 +8,7 @@ use App\Models\Consumidor;
 use App\Models\User;
 use Illuminate\Auth\Events\Registered;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Hash;
 
@@ -57,12 +58,13 @@ class ConsumidorController extends Controller
     }
     public function update(Request $request, $id)
     {
-        DB::beginTransaction();
-
-        $consumidor = Consumidor::find($id);
-        if (!$consumidor) {
-            return response()->json(['erro' => 'Usuário não encontrado'], 404);
+        $user = $request->user();
+        $consumidor = Consumidor::findOrFail($id);
+        if ($user->cannot('update', $consumidor)) {
+            abort(403);
         }
+
+        DB::beginTransaction();
         $consumidor->user()->update($request->only('nome', 'apelido', 'telefone'));
         $consumidor->user->endereco()->update($request->only(
             'rua',
@@ -77,8 +79,14 @@ class ConsumidorController extends Controller
     }
     public function destroy($id)
     {
+        $user = User::find(Auth::user()->id);
+        $consumidor = Consumidor::findOrFail($id);
+        if($user->cannot('delete', $consumidor)) {
+            abort(403);
+        }
+
         DB::beginTransaction();
-        $consumidor = Consumidor::find($id);
+        $consumidor->user->delete();
         $consumidor->delete();
         DB::commit();
         return response()->noContent();
